requirements.txt
streamlit
pandas
plotly
openai
requests
selenium
pyinstaller
python-dotenv
